move social links to footer component
