+++ src/app/modules/navbar/MobileNavbar.tsx
@@ -33,24 +33,6 @@ const links: Links = {
   // contact: '/#contact',
 }
 
-type Social = {
-  [key: string]: {
-    logo: string
-    link: string
-  }
-}
-
-const socials: Social = {
-  linkedin: {
-    logo: '/images/logos/linkedin-black.png',
-    link: 'https://www.linkedin.com/in/kevinleaves/',
-  },
-  github: {
-    logo: '/images/logos/github-mark.png',
-    link: 'https://github.com/kevinleaves',
-  },
-}
-
 export default function MobileNavbar({
   refs,
   setMenuClicked,
@@ -97,19 +79,6 @@ export default function MobileNavbar({
               </Link>
             )
           })}
-        </div>
-        <div className='flex flex-col gap-2'>
-          {Object.entries(socials).map(([key, { logo, link }]) => {
-            return (
-              <a href={link} key={key} target='blank'>
-                <Image src={logo} height={40} width={40} alt='social_logo' />
-              </a>
-            )
-          })}
-          <iframe
-            src='https://w.soundcloud.com/icon/?url=http%3A%2F%2Fsoundcloud.com%2Fkevinxle&color=black_white&size=40'
-            style={{ width: '40px', height: '40px' }}
-          ></iframe>
         </div>
       </nav>
     </motion.div>
