+++ src/app/modules/navbar/Navbar.tsx
@@ -32,24 +32,6 @@ const links: Links = {
   // contact: '/#contact',
 }
 
-type Social = {
-  [key: string]: {
-    logo: string
-    link: string
-  }
-}
-
-const socials: Social = {
-  linkedin: {
-    logo: '/images/logos/linkedin-black.png',
-    link: 'https://www.linkedin.com/in/kevinleaves/',
-  },
-  github: {
-    logo: '/images/logos/github-mark.png',
-    link: 'https://github.com/kevinleaves',
-  },
-}
-
 export default function Navbar({ refs }: Props): JSX.Element {
   const [menuClicked, setMenuClicked] = useState(false)
   const [darkModeEnabled, setDarkModeEnabled] = useDarkMode()
@@ -106,24 +88,6 @@ export default function Navbar({ refs }: Props): JSX.Element {
                     </Link>
                   )
                 })}
-              </div>
-              <div className='flex'>
-                {Object.entries(socials).map(([key, { logo, link }]) => {
-                  return (
-                    <a href={link} key={key} target='blank'>
-                      <Image
-                        src={logo}
-                        height={40}
-                        width={40}
-                        alt='social_logo'
-                      />
-                    </a>
-                  )
-                })}
-                <iframe
-                  src='https://w.soundcloud.com/icon/?url=http%3A%2F%2Fsoundcloud.com%2Fkevinxle&color=black_white&size=40'
-                  style={{ width: '40px', height: '40px' }}
-                ></iframe>
               </div>
               <div className='h-10 w-10 rounded-full border-2 border-solid border-black hover:bg-purple-500'>
                 <WiMoonAltFirstQuarter
